requirements.txt
pandas
phonenumbers
python-dateutil
pytest
